requirements.txt
anyio==3.4.0
asgiref==3.4.1
aws-xray-sdk==2.8.0
boto3==1.20.19
botocore==1.23.19
click==8.0.3
config==0.5.1
Deprecated==1.2.13
fastapi==0.70.0
future==0.18.2
h11==0.12.0
idna==3.3
jmespath==0.10.0
jsonpickle==2.0.0
postgres==4.0
psycopg2-binary==2.9.2
psycopg2-pool==1.1
pydantic==1.8.2
python-dateutil==2.8.2
redis==4.0.2
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
sse-starlette==0.9.0
starlette==0.16.0
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
wrapt==1.13.3
